+++ src/edu/jhuapl/sbmt/image2/pipelineComponents/operators/rendering/cylindricalImage/PartialCylindricalClipOperator.java
@@ -261,7 +261,7 @@ public class PartialCylindricalClipOperator extends BasePipelineOperator<vtkPoly
 
         vtkFloatArray textureCoords = new vtkFloatArray();
 
-        int numberOfPoints = polydata.GetNumberOfPoints();
+        int numberOfPoints = (int)polydata.GetNumberOfPoints();
 
         textureCoords.SetNumberOfComponents(2);
         textureCoords.SetNumberOfTuples(numberOfPoints);
